requirements.txt
# requirements.txt
PyQt6>=6.5.0
py7zr>=0.20.5          # Для .7z архивов
rarfile>=4.0            # Для .rar архивов (чтение)
pyzipper>=0.3.6         # Для паролей в ZIP архивах
Pillow>=10.0.0          # Для создания иконок